feat: make PrayFacadeService
