+++ src/main/java/com/uspray/uspray/service/HistoryService.java
@@ -6,14 +6,11 @@ import com.uspray.uspray.DTO.history.response.HistoryResponseDto;
 import com.uspray.uspray.Enums.PrayType;
 import com.uspray.uspray.domain.History;
 import com.uspray.uspray.domain.Member;
-import com.uspray.uspray.domain.Pray;
 import com.uspray.uspray.exception.ErrorStatus;
 import com.uspray.uspray.exception.model.NotFoundException;
 import com.uspray.uspray.infrastructure.HistoryRepository;
 import com.uspray.uspray.infrastructure.MemberRepository;
-import com.uspray.uspray.infrastructure.PrayRepository;
 import java.time.LocalDate;
-import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
@@ -25,15 +22,14 @@ import org.springframework.transaction.annotation.Transactional;
 @Service
 @RequiredArgsConstructor
 public class HistoryService {
-    
+
     private final HistoryRepository historyRepository;
     private final MemberRepository memberRepository;
-    private final PrayRepository prayRepository;
-    
+
     @Transactional(readOnly = true)
     public HistoryListResponseDto getHistoryList(String username, String type, int page, int size) {
         // type은 대소문자 구분하지 않습니다
-        
+
         Pageable pageable = PageRequest.of(page, size, Sort.by("deadline").descending());
         Member member = memberRepository.getMemberByUserId(username);
         Page<HistoryResponseDto> historyList;
@@ -49,12 +45,12 @@ public class HistoryService {
         return new HistoryListResponseDto(historyList.getContent(),
             historyList.getTotalPages());
     }
-    
+
     @Transactional(readOnly = true)
     public HistoryListResponseDto searchHistoryList(String username, String keyword,
         Boolean isPersonal, Boolean isShared, LocalDate startDate, LocalDate endDate, int page,
         int size) {
-        
+
         // 전체 파라미터가 null 인 경우 예외처리
         if (keyword == null && isPersonal == null && isShared == null && startDate == null
             && endDate == null) {
@@ -66,7 +62,7 @@ public class HistoryService {
             .map(HistoryResponseDto::of);
         return new HistoryListResponseDto(historyList.getContent(), historyList.getTotalPages());
     }
-    
+
     @Transactional(readOnly = true)
     public HistoryDetailResponseDto getHistoryDetail(String username, Long historyId) {
         Member member = memberRepository.getMemberByUserId(username);
@@ -77,28 +73,7 @@ public class HistoryService {
         }
         return HistoryDetailResponseDto.of(history);
     }
-    
-    @Transactional
-    public void convertPrayToHistory() {
-        List<Pray> prayList = prayRepository.findAllByDeadlineBefore(LocalDate.now());
-        for (Pray pray : prayList) {
-            History history = History.builder()
-                .pray(pray)
-                .build();
-            historyRepository.save(history);
-            prayRepository.delete(pray);
-        }
-    }
-    
-    public void createHistory(String username, Long prayId) {
-        Pray pray = prayRepository.getPrayByIdAndMemberId(prayId, username);
-        pray.complete();
-        History history = History.builder()
-            .pray(pray)
-            .build();
-        historyRepository.save(history);
-    }
-    
+
     public void deleteHistory(Long historyId, String username) {
         Member member = memberRepository.getMemberByUserId(username);
         History history = historyRepository.findByIdAndMember(historyId, member);
